fix(annotations): use Double.parseDouble for stale-session timestamps in EventGenerator and GptBotRobot
Long.parseLong fails on decimal-form timestamps serialized by SelectionExtractor, causing NumberFormatException that is silently treated as an active session — stale crashed-session annotations are never eligible for cleanup.

+++ wave/src/main/java/org/waveprotocol/box/server/robots/passive/EventGenerator.java
@@ -386,7 +386,7 @@ public class EventGenerator {
                 boolean stale = false;
                 if (parts.length >= 2 && !parts[1].isEmpty()) {
                   try {
-                    long startTimeMs = Long.parseLong(parts[1]);
+                    long startTimeMs = (long) Double.parseDouble(parts[1]);
                     stale = (now - startTimeMs > STALE_EDITING_THRESHOLD_MS);
                   } catch (NumberFormatException e) {
                     // Cannot parse startTimeMs — treat as active (safe default)
