+++ wave/src/main/java/org/waveprotocol/examples/robots/gptbot/GptBotRobot.java
@@ -321,7 +321,7 @@ public final class GptBotRobot {
         boolean stale = false;
         if (parts.length >= 2 && !parts[1].isEmpty()) {
           try {
-            long startTimeMs = Long.parseLong(parts[1]);
+            long startTimeMs = (long) Double.parseDouble(parts[1]);
             stale = (now - startTimeMs > STALE_EDITING_THRESHOLD_MS);
           } catch (NumberFormatException e) {
             // Cannot parse startTimeMs — treat as active (safe default)
